requirements.txt
absl-py==2.1.0
loguru==0.7.2
torch==2.4.0
torchvision==0.19.0
flash-attn==2.6.3
triton==3.0.0
transformer_engine @git+https://github.com/NVIDIA/TransformerEngine.git@stable
